Fixed test inheritance issues
