+++ DotnetCore/Sauce.Demo/Core.BestPractices.Web/Tests/WebTestsBase.cs
@@ -1,4 +1,6 @@
 using NUnit.Framework;
+using NUnit.Framework.Interfaces;
+using OpenQA.Selenium;
 using System;
 using System.Collections.Generic;
 
@@ -7,23 +9,20 @@ namespace Core.BestPractices.Web.Tests
     [TestFixture]
     public class WebTestsBase : AllTestsBase
     {
-        public string SauceUserName;
-        public string SauceAccessKey;
-        public Dictionary<string, object> SauceOptions;
-        public string ScreenerApiKey;
-
-        [SetUp]
-        public void Setup()
+        [TearDown]
+        public void CleanUpAfterEveryTestMethod()
         {
-            SauceUserName = Environment.GetEnvironmentVariable("SAUCE_USERNAME", EnvironmentVariableTarget.User);
-            SauceAccessKey = Environment.GetEnvironmentVariable("SAUCE_ACCESS_KEY", EnvironmentVariableTarget.User);
-            ScreenerApiKey = Environment.GetEnvironmentVariable("SCREENER_API_KEY", EnvironmentVariableTarget.User);
-            SauceOptions = new Dictionary<string, object>
-            {
-                ["username"] = SauceUserName,
-                ["accessKey"] = SauceAccessKey,
-                ["name"] = TestContext.CurrentContext.Test.Name
-            };
+            if (Driver == null)
+                return;
+            ExecuteSauceCleanupSteps();
+            Driver.Quit();
+        }
+        private void ExecuteSauceCleanupSteps()
+        {
+            var isPassed = TestContext.CurrentContext.Result.Outcome.Status
+                           == TestStatus.Passed;
+            var script = "sauce:job-result=" + (isPassed ? "passed" : "failed");
+            ((IJavaScriptExecutor)Driver).ExecuteScript(script);
         }
     }
 }
